@@ -1,5 +1,4 @@
 ﻿using System.Collections.Generic;
-using System.Linq;
 using Google.Maps.WebServices.Common;
 using Newtonsoft.Json;
 
@@ -9,8 +8,8 @@ namespace Google.Maps.WebServices.Directions
     /// <see cref="DirectionsResult" /> represents a result from the Google Maps Directions Web Service.
     /// </summary>
     /// <remarks>
-    /// See <see href="https://developers.google.com/maps/documentation/directions/overview">
-    /// Directions API</see> for more detail.
+    /// See <see href="https://developers.google.com/maps/documentation/directions/get-directions#DirectionsResponse">
+    /// DirectionsResponse</see> for more detail.
     /// </remarks>
     public class DirectionsResult
     {
@@ -23,24 +22,40 @@ namespace Google.Maps.WebServices.Directions
         /// <summary>
         /// Constructs an instance of the <see cref="DirectionsResult" /> class.
         /// </summary>
-        /// <param name="geocodedWaypoints">A collection of <see cref="GeocodedWaypoint" />.</param>
+        /// <param name="availableTravelModes">A collection of <see cref="TravelMode" />.</param>
+        /// <param name="geocodedWaypoints">A collection of <see cref="DirectionsGeocodedWaypoint" />.</param>
         /// <param name="routes">A collection of <see cref="DirectionsRoute" />.</param>
-        public DirectionsResult(List<GeocodedWaypoint> geocodedWaypoints, List<DirectionsRoute> routes)
+        public DirectionsResult(List<TravelMode> availableTravelModes, List<DirectionsGeocodedWaypoint> geocodedWaypoints, List<DirectionsRoute> routes)
         {
+            AvailableTravelModes = availableTravelModes;
             GeocodedWaypoints = geocodedWaypoints;
             Routes = routes;
         }
+
+        /// <summary>
+        /// Contains a collection of available travel modes.
+        /// </summary>
+        /// <remarks>
+        /// This field is returned when a request specifies a travel mode and gets no results.<br/>
+        /// The collection contains the available travel modes in the countries of the given set of waypoints.<br/>
+        /// This field is not returned if one or more of the waypoints are 'via waypoints'.<br/>
+        /// See <see
+        /// href="https://developers.google.com/maps/documentation/directions/get-directions#TravelMode">Geocoded
+        /// Waypoints</see> for more detail.
+        /// </remarks>
+        [JsonProperty("available_travel_modes")]
+        public List<TravelMode> AvailableTravelModes { get; } = new List<TravelMode>();
 
         /// <summary>
         /// Details about the geocoding of origin, destination, and waypoints.
         /// </summary>
         /// <remarks>
         /// See <see
-        /// href="https://developers.google.com/maps/documentation/directions/get-directions#GeocodedWaypoints">Geocoded
+        /// href="https://developers.google.com/maps/documentation/directions/get-directions#DirectionsGeocodedWaypoint">Geocoded
         /// Waypoints</see> for more detail.
         /// </remarks>
         [JsonProperty("geocoded_waypoints")]
-        public List<GeocodedWaypoint> GeocodedWaypoints { get; } = new List<GeocodedWaypoint>();
+        public List<DirectionsGeocodedWaypoint> GeocodedWaypoints { get; } = new List<DirectionsGeocodedWaypoint>();
 
         /// <summary>
         /// A collection of <see cref="DirectionsRoute" /> from the origin to the destination.
@@ -52,21 +67,5 @@ namespace Google.Maps.WebServices.Directions
         /// </remarks>
         [JsonProperty("routes")]
         public List<DirectionsRoute> Routes { get; } = new List<DirectionsRoute>();
-
-        internal static DirectionsResult Merge(DirectionsResult a, DirectionsResult b)
-        {
-            var geocodedWaypoints = a.GeocodedWaypoints.Concat(b.GeocodedWaypoints).ToList();
-            List<DirectionsRoute> directionsRoutes = new List<DirectionsRoute>();
-
-            for (int i = 0; i < a.Routes.Count || i == 0; i++)
-            {
-                for (int j = 0; j < b.Routes.Count || j == 0; j++)
-                {
-                    directionsRoutes.Add(DirectionsRoute.Merge(a.Routes[i], b.Routes[j]));
-                }
-            }
-
-            return new DirectionsResult(geocodedWaypoints, directionsRoutes);
-        }
     }
 }
